rest/requirements.txt
django==3.3.1
python==3.9
djangorestframework